refactored the main workflow
