+++ pkg/iago/feed.go
@@ -1,0 +1,6 @@
+package iago
+
+type fetcher interface {
+	readData() (feedContent, error)
+	clientName() string
+}
